Rename: admin.js => reviewer.js
